third_party/blink/web_tests/fast/canvas/check-stale-putImageData-expected.txt

CONSOLE MESSAGE: line 25: a == b? false
CONSOLE MESSAGE: line 32: a == c? false
